Power curve using Gaussian/Binomial DGPs

Generates a synthetic superpopulation via simulate_crossfit_data() and computes Monte Carlo power curves for the PP mean estimator over a grid of labeled sample sizes using simulate_power_ppi_mean().

power_curve_mean_dgp(
  n_grid,
  N,
  theta0 = 0,
  family = stats::gaussian(),
  superpop_n = 5000,
  p = 5,
  K = 5,
  seed = 1,
  alpha = 0.05,
  R = 2000
)

Arguments

n_grid

Integer vector of candidate labeled sample sizes.

N

Unlabeled sample size.

theta0

Null value for the mean estimand.

family

GLM family passed to the DGP; defaults to stats::gaussian() (with identity link) or stats::binomial() (logistic).

superpop_n

Size of the synthetic superpopulation used to estimate variance components.

p

Number of covariates in the DGP.

K

Number of folds for cross-fitting.

seed

Random seed forwarded to simulate_crossfit_data().

alpha

Two-sided significance level.

R

Number of Monte Carlo replicates passed to simulate_power_ppi_mean().

Value

Data frame with the columns produced by power_curve_mean() plus family, theta, and theta0 for downstream plotting.

Examples

power_curve_mean_dgp(
  n_grid = seq(50, 200, by = 25),
  N = 3000,
  theta0 = 0,
  family = stats::gaussian(),
  superpop_n = 6000,
  R = 1000
)
#>     n power_empirical power_exact      delta    N alpha    var_f var_res
#> 1  50           0.063  0.05069918 0.02253259 3000  0.05 11.14837 3.97508
#> 2  75           0.045  0.05102607 0.02253259 3000  0.05 11.14837 3.97508
#> 3 100           0.043  0.05133910 0.02253259 3000  0.05 11.14837 3.97508
#> 4 125           0.050  0.05163915 0.02253259 3000  0.05 11.14837 3.97508
#> 5 150           0.043  0.05192699 0.02253259 3000  0.05 11.14837 3.97508
#> 6 175           0.043  0.05220337 0.02253259 3000  0.05 11.14837 3.97508
#> 7 200           0.045  0.05246895 0.02253259 3000  0.05 11.14837 3.97508
#>     family      theta theta0
#> 1 gaussian 0.02253259      0
#> 2 gaussian 0.02253259      0
#> 3 gaussian 0.02253259      0
#> 4 gaussian 0.02253259      0
#> 5 gaussian 0.02253259      0
#> 6 gaussian 0.02253259      0
#> 7 gaussian 0.02253259      0