Type I error curve using Gaussian/Binomial DGPs

Simulates superpopulation data via simulate_crossfit_data() and evaluates empirical and analytical Type I error across an effect-size grid by repeatedly calling simulate_power_ppi_mean().

type1_error_curve_mean_dgp(
  effect_grid,
  N,
  n,
  family = stats::gaussian(),
  superpop_n = 5000,
  p = 5,
  K = 5,
  seed = 1,
  alpha = 0.05,
  R = 2000
)

Arguments

effect_grid

Numeric vector of effect sizes \(\theta - \theta_0\) to evaluate.

N

Unlabeled sample size.

n

Labeled sample size used inside simulate_power_ppi_mean().

family

GLM family passed to the DGP; defaults to stats::gaussian() (with identity link) or stats::binomial() (logistic).

superpop_n

Size of the synthetic superpopulation used to estimate variance components.

p

Number of covariates in the DGP.

K

Number of folds for cross-fitting.

seed

Random seed forwarded to simulate_crossfit_data().

alpha

Two-sided significance level.

R

Number of Monte Carlo replicates passed to simulate_power_ppi_mean().

Value

Data frame with columns effect_size, type1_empirical, type1_exact, theta, theta0, and additional metadata mirroring the inputs.

Examples

type1_error_curve_mean_dgp(
  effect_grid = seq(-0.4, 0.4, by = 0.05),
  N = 4000,
  n = 200,
  family = stats::gaussian(),
  R = 2000
)
#>    effect_size type1_empirical type1_exact      theta      theta0    N   n
#> 1        -0.40          0.7630  0.74039583 0.02514856  0.42514856 4000 200
#> 2        -0.35          0.6365  0.62514539 0.02514856  0.37514856 4000 200
#> 3        -0.30          0.4945  0.49742363 0.02514856  0.32514856 4000 200
#> 4        -0.25          0.3520  0.37005949 0.02514856  0.27514856 4000 200
#> 5        -0.20          0.2550  0.25591729 0.02514856  0.22514856 4000 200
#> 6        -0.15          0.1395  0.16439660 0.02514856  0.17514856 4000 200
#> 7        -0.10          0.0760  0.09980805 0.02514856  0.12514856 4000 200
#> 8        -0.05          0.0420  0.06222849 0.02514856  0.07514856 4000 200
#> 9         0.00          0.0450  0.05000000 0.02514856  0.02514856 4000 200
#> 10        0.05          0.0465  0.06222849 0.02514856 -0.02485144 4000 200
#> 11        0.10          0.0925  0.09980805 0.02514856 -0.07485144 4000 200
#> 12        0.15          0.1465  0.16439660 0.02514856 -0.12485144 4000 200
#> 13        0.20          0.2350  0.25591729 0.02514856 -0.17485144 4000 200
#> 14        0.25          0.3760  0.37005949 0.02514856 -0.22485144 4000 200
#> 15        0.30          0.4835  0.49742363 0.02514856 -0.27485144 4000 200
#> 16        0.35          0.6490  0.62514539 0.02514856 -0.32485144 4000 200
#> 17        0.40          0.7600  0.74039583 0.02514856 -0.37485144 4000 200
#>    alpha    var_f var_res   family
#> 1   0.05 10.77206  4.1787 gaussian
#> 2   0.05 10.77206  4.1787 gaussian
#> 3   0.05 10.77206  4.1787 gaussian
#> 4   0.05 10.77206  4.1787 gaussian
#> 5   0.05 10.77206  4.1787 gaussian
#> 6   0.05 10.77206  4.1787 gaussian
#> 7   0.05 10.77206  4.1787 gaussian
#> 8   0.05 10.77206  4.1787 gaussian
#> 9   0.05 10.77206  4.1787 gaussian
#> 10  0.05 10.77206  4.1787 gaussian
#> 11  0.05 10.77206  4.1787 gaussian
#> 12  0.05 10.77206  4.1787 gaussian
#> 13  0.05 10.77206  4.1787 gaussian
#> 14  0.05 10.77206  4.1787 gaussian
#> 15  0.05 10.77206  4.1787 gaussian
#> 16  0.05 10.77206  4.1787 gaussian
#> 17  0.05 10.77206  4.1787 gaussian